Chunk Update
Amended the way the chunks are stored. Now each chunk is stored in a path like this:
\resources\chunks\difficulty\'difficulty_number.txt'

With this, moved the existing chunks

Added splitter.py into resources/chunks
splitter.py splits a single file with chunks into separate ones in a folder. This is to make patching and modding the game more convenient, as all the chunks of the same difficulty can now be modified from a single file

Edited chunks.py:
redo_chunk function to accept the difficulty of the chunk and the path to the chunk rather than it's name, and to be able to redo all the chunks in the difficulty directory

got rid of the '_refactored.txt' suffix for chunks translated by the redo_chunk, now it just overwrites the chunk

Edited tower.py:
load_chunk now accepth the path to the chunk
replaced get_chunk_name with get_chunk_path which returns the path to the chunk, and also doesn't rely on random.randint to generate the number of the chunk, instead choosing a random chunk from the difficulty directory